+++ src/lib/stores/projectStores.js
@@ -4,7 +4,7 @@ import { writable } from 'svelte/store';
 // within a ProjectInfoModal. Every component that will update the project
 // inside the store, will also update the to-be-displayed project in a modal.
 /** @type {import('svelte/types/runtime/store').Writable<import('$lib/types').Project|undefined>}} */
-export const modalProject = writable(undefined);
+export const projectInfoModal = writable(undefined);
 
 // Context project store
 export const contextProject = writable({